requirements.txt
# Minimal API Testing Requirements

# Build essentials first
wheel==0.43.0
numpy==1.26.4  # for kiwipiepy build

# Core LangGraph and LangChain
langgraph==1.0.4
langchain-core==1.1.0

# API integrations
langchain-openai==1.1.0
langchain-google-genai==3.2.0
langchain-anthropic==1.2.0
langchain-upstage==0.7.5
langchain-perplexity==1.1.0
langchain-mistralai==1.1.0
langchain-groq==1.1.0
langchain-cohere==0.5.0

# Environment and utilities
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
ipython==8.28.0  # required by langchain-teddynote

# Web API (Streamlit for local UI)
fastapi==0.122.0
uvicorn[standard]==0.38.0
requests==2.32.5
streamlit==1.51.0  # local UI; exclude from production if not needed
email-validator==2.1.0.post1  # for Pydantic EmailStr

# Known Windows wheel requirement (ASCII comments only)
PyYAML==6.0.1

# Additional dependencies
tokenizers==0.20.3
# langchain-upstage 0.7.5 requires pypdf<5.0.0
pypdf==4.3.1
